requirements.txt
# DevOps Python Automation - Required Packages
# Course: Week 3-4 Python Automation
# Python Version: 3.9+

# HTTP & API Libraries
requests==2.31.0

# Configuration & Data Formats
PyYAML==6.0.1
python-dotenv==1.0.0

# Google Cloud Platform SDKs
google-cloud-storage==2.14.0
google-cloud-compute==1.15.0
google-cloud-bigquery==3.14.1

# Container & Orchestration
docker==7.0.0
kubernetes==28.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# CLI Tools
click==8.1.7

# Terminal Output
rich==13.7.0
colorama==0.4.6
tabulate==0.9.0

# Utilities
Jinja2==3.1.2
psutil==5.9.8

# Code Quality (Optional but Recommended)
pylint==3.0.3
flake8==6.1.0
black==23.12.1
mypy==1.7.1

# Type Stubs for Better Type Checking
types-requests==2.31.0.10
types-PyYAML==6.0.12.12
